requirements.txt
torch
torchvision
torchaudio
opencv-python
pillow
tqdm
numpy
matplotlib
scikit-learn
pandas
scipy
git+https://github.com/openai/CLIP.git
